examples/robot-template/requirements.txt
paho-mqtt>=1.6.0,<2.0.0
pyyaml>=6.0
requests>=2.28.0
